Edit

+++ app/models.py
@@ -420,7 +420,7 @@ class Classifier(db.Model):
     description = db.Column(db.String(512), index=False)
     region = db.Column(db.String(64), index=True)
     active = db.Column(db.Boolean, default=True, index=True)
-    threshold = db.Column(db.Integer, index=False)
+    threshold = db.Column(db.Float, index=False)
     surveys = db.relationship('Survey', backref='classifier', lazy='dynamic')
 
     def __repr__(self):
